+++ plugins/textshape/tests/TestDocumentLayout.cpp
@@ -282,7 +282,7 @@ void TestDocumentLayout::testAdvancedLineSpacing()
     QVERIFY(block.isValid());
     m_blockLayout = block.layout();
     // qDebug() << m_blockLayout->lineAt(0).y();
-    QVERIFY(qAbs(m_blockLayout->lineAt(0).y() - (77.6 + qMax(12 * 1.2, 5.0))) < ROUNDING); // 92
+    QVERIFY(qAbs(m_blockLayout->lineAt(0).y() - (77.6 + qMax(qreal(12.0 * 1.2), qreal(5.0)))) < ROUNDING); // 92
     block = block.next(); // line6
     QVERIFY(block.isValid());
     m_blockLayout = block.layout();
